rename view_part_system.rs files to visualizer_system.rs files
