+++ crates/re_space_view_bar_chart/src/lib.rs
@@ -3,6 +3,6 @@
 //! A Space View that shows a single bar chart.
 
 mod space_view_class;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::BarChartSpaceView;
