+++ crates/re_space_view_bar_chart/src/visualizer_system.rs
