+++ crates/re_space_view_dataframe/src/lib.rs
@@ -3,6 +3,6 @@
 //! A Space View that shows the data contained in entities in a table.
 
 mod space_view_class;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::DataframeSpaceView;
