+++ crates/re_space_view_tensor/src/lib.rs
@@ -6,6 +6,6 @@ mod dimension_mapping;
 mod space_view_class;
 mod tensor_dimension_mapper;
 mod tensor_slice_to_gpu;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::TensorSpaceView;
