+++ crates/re_space_view_tensor/src/visualizer_system.rs
