+++ crates/re_space_view_text_document/src/lib.rs
@@ -3,6 +3,6 @@
 //! A simple Space View that shows a single text document.
 
 mod space_view_class;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::TextDocumentSpaceView;
