+++ crates/re_space_view_text_document/src/visualizer_system.rs
