+++ crates/re_space_view_text_log/src/lib.rs
@@ -3,6 +3,6 @@
 //! A Space View that shows `TextLog` entries in a table and scrolls with the active time.
 
 mod space_view_class;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::TextSpaceView;
