+++ crates/re_space_view_text_log/src/space_view_class.rs
@@ -10,7 +10,7 @@ use re_viewer_context::{
     ViewQuery, ViewerContext,
 };
 
-use super::view_part_system::{Entry, TextLogSystem};
+use super::visualizer_system::{Entry, TextLogSystem};
 
 // TODO(andreas): This should be a blueprint component.
 #[derive(Clone, PartialEq, Eq, Default)]
