+++ crates/re_space_view_text_log/src/visualizer_system.rs
