+++ crates/re_space_view_time_series/src/lib.rs
@@ -3,6 +3,6 @@
 //! A Space View that shows plots over Rerun timelines.
 
 mod space_view_class;
-mod view_part_system;
+mod visualizer_system;
 
 pub use space_view_class::TimeSeriesSpaceView;
