+++ crates/re_space_view_time_series/src/space_view_class.rs
@@ -12,7 +12,7 @@ use re_viewer_context::{
     SpaceViewSystemExecutionError, SystemExecutionOutput, ViewQuery, ViewerContext,
 };
 
-use crate::view_part_system::{PlotSeriesKind, TimeSeriesSystem};
+use crate::visualizer_system::{PlotSeriesKind, TimeSeriesSystem};
 
 #[derive(Clone, Default)]
 pub struct TimeSeriesSpaceViewState {
