+++ examples/rust/custom_space_view/src/color_coordinates_visualizer_system.rs
